refac: @JvmField 제거

+++ backend/src/main/java/com/team05/petmeeting/domain/cheer/dto/CheerStatusDto.kt
@@ -1,7 +1,7 @@
 package com.team05.petmeeting.domain.cheer.dto
 
 data class CheerStatusDto(
-    val usedToday: Long,
+    val usedToday: Int,
     val remainingToday: Int,
     val resetAt: String // "2024-08-10T00:00:00"
 )
